데이터 갱신 및 스크립트 수정
1. 데이터베이스 새로운 파일로 갱신 (수소~탄소)
2. roomTempPos가 int에서 float으로 변경됨에 따라
이를 사용하는 다른 클래스의 변수형을 float으로 변경.

+++ Assets/StatScreenFolder/Scripts/ThermoBar.cs
@@ -14,12 +14,17 @@ public class ThermoBar : MonoBehaviour {
     public int valSolid;
     public int valLiquid;
     public int valGas;
-    public int pointerVal;
+
+	//2016-03-19 Changed int -> float
+    public float pointerVal;
+	//
     public int barSolid;
     public int barLiquid;
     public int barGas;
     public ChemicalStates currentCState;
-    public int currentCStateValue;
+    
+	//2016-03-19 Changed int -> float
+	public float currentCStateValue;
 
     // Use this for initialization
     void Start () {
